+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxInfo.hs
@@ -59,6 +59,7 @@ module Cardano.Ledger.Alonzo.TxInfo
     languages,
     -- DEPRECATED
     validPlutusdata,
+    getTxOutDatum,
   )
 where
 
@@ -67,8 +68,7 @@ where
 import Cardano.Binary (FromCBOR (..), ToCBOR (..), decodeFull')
 import Cardano.Crypto.Hash.Class (Hash, hashToBytes)
 import Cardano.Ledger.Address (Addr (..), RewardAcnt (..))
-import Cardano.Ledger.Alonzo.Data (Data (..), Datum (..), getPlutusData)
-import Cardano.Ledger.Alonzo.Language (Language (..))
+import Cardano.Ledger.Alonzo.Data (Data (..), Datum, getPlutusData)
 import Cardano.Ledger.Alonzo.Scripts
   ( AlonzoScript (..),
     ExUnits (..),
@@ -102,9 +102,9 @@ import Cardano.Ledger.Credential
   )
 import qualified Cardano.Ledger.Crypto as CC (Crypto)
 import Cardano.Ledger.Keys (KeyHash (..), hashKey)
+import Cardano.Ledger.Language (Language (..))
 import Cardano.Ledger.Mary.Value (AssetName (..), MaryValue (..), MultiAsset (..), PolicyID (..))
 import Cardano.Ledger.SafeHash (SafeHash, extractHash, hashAnnotated)
-import Cardano.Ledger.Serialization (Sized (sizedValue))
 import qualified Cardano.Ledger.Shelley.HardForks as HardForks
 import Cardano.Ledger.Shelley.TxBody
   ( DCert (..),
@@ -483,17 +483,9 @@ class ExtendedUTxO era where
     ScriptPurpose (EraCrypto era) ->
     Maybe (Data era)
 
-  getTxOutDatum ::
-    TxOut era -> Datum era
-
-  allOuts ::
-    TxBody era ->
-    [TxOut era]
-  allOuts = map sizedValue . allSizedOuts
-
-  allSizedOuts ::
-    TxBody era ->
-    [Sized (TxOut era)]
+getTxOutDatum :: AlonzoEraTxOut era => TxOut era -> Datum era
+getTxOutDatum txOut = txOut ^. datumTxOutF
+{-# DEPRECATED getTxOutDatum "In favor of `datumTxOutF`" #-}
 
 alonzoTxInfo ::
   forall era.