+++ CHANGELOG.md
@@ -13,7 +13,10 @@ in the naming of release branches.
 
 - Start on the `cardano-ledger-api` package and implement
   `setMinCoinTxOut`+`setMinCoinSizedTxOut`: #2995
-- Added `getMinCoinTxOut`/`getMinCoinSizedTxOut`: #3008
+- Added `getMinCoinTxOut`/`getMinCoinSizedTxOut` to `EraTxOut`: #3008
+- Added `getMinFeeTx` to `EraTx`
+- Added `datumTxOutF` to `AlonzoEraTxOut`
+- Added `allInputsTxBodyF`
 
 ### Changed
 - Changed `mint` field type to `MultiAsset (Crypto era)` in `MATxBody`, `AlonzoTxBody`, `BabbageTxBody`
@@ -25,8 +28,12 @@ in the naming of release branches.
 
 - Deprecated the `validPlutusdata` function: #3006
 - Deprecated the misspelled `HasAlgorithm` type alias: #3007
-- Deprecate `evaluateMinLovelaceOutput` in favor of newly added
-  `getMinCoinTxOut`/`getMinCoinSizedTxOut` #3008
+- Deprecated `CLI.evaluateMinLovelaceOutput` in favor of newly added
+  `EraTxOut.getMinCoinTxOut`/`EraTxOut.getMinCoinSizedTxOut` #3008
+- Deprecated `minfee` and `CLI.evaluateMinfee` in favor of new `EraTx.getMinFeeTx`
+- Deprecated `ExtendedUTxO.getTxOutDatum` in favor of new `AlonzoEraTxOut.datumTxOutF`
+- Removed `ExtendedUTxO.allOuts` and `ExtendedUTxO.allSizedOuts` in favor of
+  `BabbageEraTxBody.allInputsTxBodyF`
 
 ## Release branch 1.1.x